/* -----Initial MasterMind game----- */
The game generates a random code that the user will enter when ask.
Once the user code is entered, the game will compare the random code and the user input code.
If the user enter the correct code the user will be notified,
else the game will notify him which number (exist but in the wrong spot, does not exist).
The user haves 8 tries before game over.

-----------------what's new------------------

-2 files created :Test.java and MasterMind.java
-3 methods created : MasterMindGame(), randomCode(), getUserCode()
-validation method is till in progress